cont

+++ src/main/java/com/github/lindenb/jvarkit/tools/bam2svg/WesCnvSvg.java
@@ -529,7 +529,7 @@ public class WesCnvSvg  extends Launcher {
 			w.writeStartElement("g");
 			w.writeAttribute("class", "maing");
 			
-			int y=0;
+			double y=0;
 			w.writeStartElement("g");
 			w.writeComment("interval background");
 			for(final CaptureInterval ci:userIntervals)
